pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devbot"
version = "0.1.0"
description = "Discord AI development assistant bot"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "anthropic>=0.40",
    "openai>=1.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "platformdirs>=4.0",
    "python-dotenv>=1.0",
]

[project.scripts]
devbot = "devbot.cli:cli_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["devbot*"]

[tool.setuptools.package-data]
"devbot.workflow" = ["*.yaml"]
